refactor(cellbuf): rename grid to screen and clean up
