+++ cellbuf/method.go
@@ -1,11 +1,11 @@
 package cellbuf
 
-// WidthMethod is a type that represents the how the renderer should calculate
-// the display width of cells.
-type WidthMethod uint8
+// Method is a type that represents the how the renderer should calculate the
+// display width of cells.
+type Method uint8
 
 // Display width modes.
 const (
-	WcWidth WidthMethod = iota
+	WcWidth Method = iota
 	GraphemeWidth
 )
